Rewrite the tilemap importer
Tiled importer now imports collision data to a tilemap, and image data to a sprite. Image is blitted from all the tiles and joined into one.
This simplifies the workflow for map editing, you should only really need to change things in Tiled now.
Also changed the object container to YSort. Not sure if the entire map can be in Ysort, I keep node2D for now.