Removed most tree objects
Removed most of the tree objects on the map. Godot cannot handle that many objects in editor, they would need to be loaded as separate scenes. Currently we dont have a system for that but for now I suggest we keep it as is and re-add the trees later as unpassable terrain.